core(explore): rm const and move to demojs
